requirements.txt
openai_agents
python-dotenv
chainlit
streamlit
pydantic
